implemented softkill via xclientmsg

+++ src/workspace/mod.rs
@@ -5,7 +5,11 @@ pub mod workspace_navigation;
 use self::workspace_layout::WorkspaceLayout;
 
 use crate::{
-    config::Config, screeninfo::ScreenSize, windowmanager::movement::Movement,
+    atom::Atom,
+    auxiliary::{atom_name, get_internal_atom},
+    config::Config,
+    screeninfo::ScreenSize,
+    windowmanager::movement::Movement,
     windowstate::WindowState,
 };
 
@@ -209,15 +213,72 @@ impl Workspace {
         }
     }
 
-    pub fn kill_window(&mut self, winid: &u32) {
-        //TODO implement soft kill via client message over x
-        //(Tell window to close itself)
-        //https://github.com/DHBW-FN/OxideWM/issues/46
-        self.remove_window(winid);
-        self.connection
-            .kill_client(*winid)
-            .expect("Could not kill client");
+    fn is_softkill_supported(&self, winid: u32) -> bool {
+        let atom_id = get_internal_atom(&self.connection, Atom::WmProtocols.as_ref());
+
         self.connection.flush().unwrap();
+        let atom_reply = self
+            .connection
+            .get_property(false, winid, atom_id, AtomEnum::ANY, 0, 1024)
+            .unwrap()
+            .reply();
+
+        if let Ok(atom_reply) = atom_reply {
+            self.connection.flush().unwrap();
+
+            let prop_type = match atom_reply.type_ {
+                0 => return false, // Null response
+                atomid => atom_name(&self.connection, atomid),
+            };
+
+            if prop_type == "ATOM" {
+                let atoms = atom_reply
+                    .value32()
+                    .unwrap()
+                    .map(|a| atom_name(&self.connection, a))
+                    .collect::<Vec<String>>();
+
+                return atoms.iter().any(|p| p == "WM_DELETE_WINDOW");
+            }
+        }
+        false
+    }
+
+    pub fn kill_window(&mut self, winid: &u32) {
+        if self.is_softkill_supported(*winid) {
+            self.execute_softkill(*winid);
+        } else {
+            self.connection
+                .kill_client(*winid)
+                .expect("Could not kill client");
+            self.remove_window(winid);
+        }
+        self.connection.flush().unwrap();
+    }
+
+    fn execute_softkill(&mut self, winid: u32) {
+        let type_ = get_internal_atom(&self.connection, Atom::WmProtocols.as_ref());
+
+        let data = ClientMessageData::from([
+            get_internal_atom(&self.connection, Atom::WmDeleteWindow.as_ref()),
+            0,
+            0,
+            0,
+            0,
+        ]);
+
+        let clnt_msg_vnt = ClientMessageEvent {
+            response_type: CLIENT_MESSAGE_EVENT,
+            format: 32,
+            sequence: 0,
+            window: winid,
+            type_,
+            data,
+        };
+
+        self.connection
+            .send_event(false, winid, EventMask::NO_EVENT, clnt_msg_vnt)
+            .unwrap();
     }
 
     pub fn remove_window(&mut self, win_id: &u32) {
